// src/generated/WorkspaceUsageResult.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { U64String } from "./U64String";
import type { WorkspaceUsageEntry } from "./WorkspaceUsageEntry";

export type WorkspaceUsageResult = { missions: Array<WorkspaceUsageEntry>, total_bytes: U64String, };