// src/generated/MissionControlParams.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { Id } from "./Id";
import type { MissionControlAction } from "./MissionControlAction";
import type { U64String } from "./U64String";

export type MissionControlParams = { request_id: Id, mission_id: Id, expected_revision: U64String, action: MissionControlAction, };