// src/generated/MutationResult.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { Id } from "./Id";
import type { U64String } from "./U64String";

export type MutationResult = { mission_id: Id, revision: U64String, event_seq: U64String, entity_ids: Array<Id>, };